ketten/requirements.txt
Django==1.4.3
djangorestframework==2.2.0
